fix: :bug: Fix bug with db not having the server
`RepeatplayButton.callback` and `regex()` both had the same issue: they tried to write to the database, while they were assuming that the server was in the db, which may not be the case. If the db is (for example) {} (empty), then setting db[server][key] will result in an error, since the server is not in the db. Other commands already fixed this. This fix checks if the server is in the db, and if it isn't it adds it to the db.

Resolves https://koviubi56.statuspage.io/incidents/5rdgbp0cg1vd