fix: :bug: Fix 429
Most 429 errors on REPLIT can be mitigated by running "kill 1", because it gives a new IP address. So this commit adds a try-except to client.run, and if there's a discord.errors.HTTPException (and env:KEEPALIVE == 1), we call kill 1. But (I.) Discord might have banned our bot, or (II.) the HTTPException might not be 429. In that case :(